+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to Merlin will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.13.0b2]
+### Added
+- Ability to turn off the auto-restart functionality of the monitor with `--no-restart`
+- Tests for the monitor files
+
+### Changed
+- Refactored the `main.py` module so that it's broken into smaller, more-manageable pieces
+
 ## [1.13.0b1]
 ### Added
 - API documentation for Merlin's core codebase